support methods on named slice types

+++ tests/XFAIL/generics_basic/main.rs
@@ -65,7 +65,7 @@ impl Unknown {
 
 pub fn map_keys(m: Rc<RefCell<Option<BTreeMap<K, Rc<RefCell<Option<V>>>>>>>) -> Rc<RefCell<Option<Vec<K>>>> {
 
-    let mut r = Rc::new(RefCell::new(Some(Vec::with_capacity((*m.borrow().as_ref().unwrap()).len()))));
+    let mut r = Rc::new(RefCell::new(Some(Vec::with_capacity(((*m.borrow().as_ref().unwrap()).len()) as usize))));
     for (k, _) in (*m.borrow().as_ref().unwrap()).clone() {
         {(*r.borrow_mut()).get_or_insert_with(Vec::new).push(k); r.clone()};
     }
