+++ tests/closures_basic/main.rs
@@ -45,7 +45,7 @@ pub fn make_adder(x: Rc<RefCell<Option<i32>>>) -> Rc<RefCell<Option<Box<dyn Fn(R
 
 pub fn apply_operation(nums: Rc<RefCell<Option<Vec<i32>>>>, op: Rc<RefCell<Option<Box<dyn Fn(Rc<RefCell<Option<i32>>>) -> Rc<RefCell<Option<i32>>>>>>>) -> Rc<RefCell<Option<Vec<i32>>>> {
 
-    let mut result = Rc::new(RefCell::new(Some(vec![0; (*nums.borrow().as_ref().unwrap()).len()])));
+    let mut result = Rc::new(RefCell::new(Some(vec![0; ((*nums.borrow().as_ref().unwrap()).len()) as usize])));
     { let __range_guard = nums.borrow(); let __range_values = __range_guard.as_ref().map(|__v| __v.as_slice()).unwrap_or(&[]); for (i, num) in __range_values.iter().copied().enumerate() {
         (*result.borrow_mut().as_mut().unwrap())[i] = (*{ let __f_guard = op.borrow(); let __f = __f_guard.as_ref().unwrap(); (*__f)(Rc::new(RefCell::new(Some(num)))) }.borrow().as_ref().unwrap());
     } }
