+++ tests/slice_operations_advanced/main.rs
@@ -25,7 +25,7 @@ where
 
 fn main() {
         // Slice with capacity
-    let mut s = Rc::new(RefCell::new(Some({ let mut v = Vec::with_capacity(10 as usize); v.resize(3 as usize, 0); v })));
+    let mut s = Rc::new(RefCell::new(Some({ let mut v = Vec::with_capacity((10) as usize); v.resize((3) as usize, 0); v })));
     print!("len={} cap={} {}\n", (*s.borrow().as_ref().unwrap()).len(), (*s.borrow().as_ref().unwrap()).capacity(), format_slice(&s));
 
         // Append beyond capacity
@@ -33,11 +33,11 @@ fn main() {
     print!("len={} cap={} {}\n", (*s.borrow().as_ref().unwrap()).len(), (*s.borrow().as_ref().unwrap()).capacity(), format_slice(&s));
 
         // Three-index slice
-    let mut s2 = Rc::new(RefCell::new(Some({ let __seq = { let __seq_holder = s.clone(); let __seq_guard = __seq_holder.borrow(); let __cloned = (*__seq_guard.as_ref().unwrap()).clone(); drop(__seq_guard); __cloned }; let _slice = &__seq[2 as usize..5 as usize]; let mut _v = Vec::with_capacity((7 - 2) as usize); _v.extend_from_slice(_slice); _v })));
+    let mut s2 = Rc::new(RefCell::new(Some({ let __seq = { let __seq_holder = s.clone(); let __seq_guard = __seq_holder.borrow(); let __cloned = (*__seq_guard.as_ref().unwrap()).clone(); drop(__seq_guard); __cloned }; let _slice = &__seq[(2) as usize..(5) as usize]; let mut _v = Vec::with_capacity(((7) - (2)) as usize); _v.extend_from_slice(_slice); _v })));
     print!("s2: len={} cap={} {}\n", (*s2.borrow().as_ref().unwrap()).len(), (*s2.borrow().as_ref().unwrap()).capacity(), format_slice(&s2));
 
         // Copy
-    let mut s3 = Rc::new(RefCell::new(Some(vec![0; 3])));
+    let mut s3 = Rc::new(RefCell::new(Some(vec![0; (3) as usize])));
     let mut n = { let _src = ((*s.borrow().as_ref().unwrap())).clone(); let _n = std::cmp::min(((*s3.borrow().as_ref().unwrap())).len(), _src.len()); for _i in 0.._n { (*s3.borrow_mut().as_mut().unwrap())[_i] = _src[_i].clone(); } Rc::new(RefCell::new(Some(_n as i32))) };
     print!("Copied {} elements: {}\n", { let __v = (*n.borrow().as_ref().unwrap()).clone(); __v }, format_slice(&s3));
